% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coedit_text.R
\name{coedit_text}
\alias{coedit_text}
\title{Run the CoeditorAI RStudio Addin on the currently selected text}
\usage{
coedit_text(original_text = get_selection(), debug = FALSE)
}
\arguments{
\item{original_text}{A character. Edited text. If a text is not provided,
the current text selection from the active file in RStudio IDE
is used by default.}

\item{debug}{A logical. The \code{TRUE} function returns a shiny app object
of class \code{shiny.appobj} for debugging and testing purposes
without actually running the app.
By default, this option is set to \code{FALSE}.}
}
\value{
A character string.
The last output text generated by the CoeditorAI Add-in session.
}
\description{
Run the CoeditorAI RStudio Addin on the currently selected text
}
\examples{
if (interactive()) {
  "Select text in RStudio IDE or pass it as a character string" |>
  coedit_text()
}

if (interactive()) {
  coedit_text()
}
}
