% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoeditorAI-package.R
\docType{package}
\name{CoeditorAI-package}
\alias{CoeditorAI}
\alias{CoeditorAI-package}
\title{CoeditorAI: Write Analytical Texts Faster that Read Better}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'CoeditorAI' is a plugin that provides Artificial Intelligence (AI) / Large Language Models (LLM) support, allowing you to write and edit professional content faster and directly inside the 'RStudio IDE'. You can rewrite, proofread, and translate text while coding, all in one place. This plugin is perfect for Data Scientists who don't like wasting time on text editing and want to work faster, write better, and communicate their findings effectively.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://coeditorai.com}
  \item Report bugs at \url{https://github.com/CoeditorAI/CoeditorAI-RStudio/issues}
}

}
\author{
\strong{Maintainer}: Kamil Wais \email{kamil.wais@gmail.com}

Authors:
\itemize{
  \item Olesia Wais \email{olesia.wais@gmail.com}
}

Other contributors:
\itemize{
  \item Kamil Wais InnowacjeSystemowe.pl [copyright holder, funder]
}

}
\keyword{internal}
